Crate proc_macro

Unstable (proc_macro_lib)

A support library for macro authors when defining new macros.

This library, provided by the standard distribution, provides the types consumed in the interfaces of procedurally defined macro definitions. Currently the primary use of this crate is to provide the ability to define new custom derive modes through #[proc_macro_derive].

Added recently as part of RFC 1681 this crate is currently unstable and requires the #![feature(proc_macro_lib)] directive to use.

Note that this crate is intentionally very bare-bones currently. The main type, TokenStream, only supports fmt::Display and FromStr implementations, indicating that it can only go to and come from a string. This functionality is intended to be expanded over time as more surface area for macro authors is stabilized.

Structs

LexError 	 [Unstable]
Error returned from TokenStream::from_str.

TokenStream 	 [Unstable]
The main type provided by this crate, representing an abstract stream of tokens.