Module proc_macro::bridge::client

ğŸ”¬This is a nightly-only experimental API. (proc_macro_internals #27812)

Client-side types.

Macros

define_client_side ğŸ”’ Experimental

define_handles ğŸ”’ Experimental

Structs

Bridge ğŸ”’ Experimental

ClientExperimental

A client-side RPC entry-point, which may be using a different proc_macro from the one used by the server, but can be invoked compatibly.

FreeFunctions ğŸ”’ Experimental

HandleCountersExperimental

HandleStore ğŸ”’ Experimental

SourceFile ğŸ”’ Experimental

Span ğŸ”’ Experimental

TokenStream ğŸ”’ Experimental

Enums

BridgeState ğŸ”’ Experimental

BridgeStateL ğŸ”’ Experimental

ProcMacroExperimental

Constants

BRIDGE_STATE ğŸ”’ Experimental

Functions

is_available ğŸ”’ Experimental

maybe_install_panic_hook ğŸ”’ Experimental

run_client ğŸ”’ Experimental

Client-side helper for handling client panics, entering the bridge, deserializing input and serializing output.