Module proc_macro::bridge::server

🔬This is a nightly-only experimental API. (proc_macro_internals #27812)

Server-side traits.

Macros

associated_fn 🔒 Experimental

Declare an associated fn of one of the traits below, adding necessary default bodies.

declare_server_traits 🔒 Experimental

define_dispatcher_impl 🔒 Experimental

define_mark_types_impls 🔒 Experimental

Structs

CrossThreadExperimental

Dispatcher 🔒 Experimental

MarkedTypes 🔒 Experimental

MaybeCrossThreadExperimental

RunningSameThreadGuard 🔒 Experimental

Keep ALREADY_RUNNING_SAME_THREAD (see also its documentation) set to true, preventing same-thread reentrance.

SameThreadExperimental

Constants

ALREADY_RUNNING_SAME_THREAD 🔒 Experimental

While running a proc-macro with the same-thread executor, this flag will be set, forcing nested proc-macro invocations (e.g. due to TokenStream::expand_expr) to be run using a cross-thread executor.

Traits

DispatcherTraitExperimental

ExecutionStrategyExperimental

FreeFunctionsExperimental

MessagePipeExperimental

A message pipe used for communicating between server and client threads.

ServerExperimental

SourceFileExperimental

SpanExperimental

SymbolExperimental

TokenStreamExperimental

TypesExperimental

Functions

run_server 🔒 Experimental