Trait proc_macro::bridge::Mark

trait Mark {
    type Unmarked;

    // Required method
    fn mark(unmarked: Self::Unmarked) -> Self;
}

🔬This is a nightly-only experimental API. (proc_macro_internals #27812)

Helper to wrap associated types to allow trait impl dispatch. That is, normally a pair of impls for T::Foo and T::Bar can overlap, but if the impls are, instead, on types like Marked<T::Foo, Foo> and Marked<T::Bar, Bar>, they can’t.

Required Associated Types

type Unmarked

🔬This is a nightly-only experimental API. (proc_macro_internals #27812)

Required Methods

fn mark(unmarked: Self::Unmarked) -> Self

🔬This is a nightly-only experimental API. (proc_macro_internals #27812)