proc_macro

Enum EscapeError

#[non_exhaustive]
pub enum EscapeError {

    ZeroChars,
    MoreThanOneChar,
    LoneSlash,
    InvalidEscape,
    BareCarriageReturn,
    BareCarriageReturnInRawString,
    EscapeOnlyChar,
    TooShortHexEscape,
    InvalidCharInHexEscape,
    OutOfRangeHexEscape,
    NoBraceInUnicodeEscape,
    InvalidCharInUnicodeEscape,
    EmptyUnicodeEscape,
    UnclosedUnicodeEscape,
    LeadingUnderscoreUnicodeEscape,
    OverlongUnicodeEscape,
    LoneSurrogateUnicodeEscape,
    OutOfRangeUnicodeEscape,
    UnicodeEscapeInByte,
    NonAsciiCharInByte,
    NulInCStr,
    UnskippedWhitespaceWarning,
    MultipleSkippedLinesWarning,

}

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Mostly relating to malformed escape sequences, but also a few other problems.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ZeroChars

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Expected 1 char, but 0 were found.

MoreThanOneChar

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Expected 1 char, but more than 1 were found.

LoneSlash

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Escaped ’' character without continuation.

InvalidEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Invalid escape character (e.g. ‘\z’).

BareCarriageReturn

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Raw ‘\r’ encountered.

BareCarriageReturnInRawString

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Raw ‘\r’ encountered in raw string.

EscapeOnlyChar

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Unescaped character that was expected to be escaped (e.g. raw ‘\t’).

TooShortHexEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Numeric character escape is too short (e.g. ‘\x1’).

InvalidCharInHexEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Invalid character in numeric escape (e.g. ‘\xz’)

OutOfRangeHexEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Character code in numeric escape is non-ascii (e.g. ‘\xFF’).

NoBraceInUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

‘\u’ not followed by ‘{’.

InvalidCharInUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Non-hexadecimal value in ‘\u{..}’.

EmptyUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

‘\u{}’

UnclosedUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

No closing brace in ‘\u{..}’, e.g. ‘\u{12’.

LeadingUnderscoreUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

‘\u{_12}’

OverlongUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

More than 6 characters in ‘\u{..}’, e.g. ‘\u{10FFFF_FF}’

LoneSurrogateUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Invalid in-bound unicode character code, e.g. ‘\u{DFFF}’.

OutOfRangeUnicodeEscape

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Out of bounds unicode character code, e.g. ‘\u{FFFFFF}’.

UnicodeEscapeInByte

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Unicode escape code in byte literal.

NonAsciiCharInByte

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

Non-ascii character in byte literal, byte string literal, or raw byte string literal.

NulInCStr

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

\0 in a C string literal.

UnskippedWhitespaceWarning

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

After a line ending with ’', the next line contains whitespace characters that are not skipped.

MultipleSkippedLinesWarning

🔬This is a nightly-only experimental API. (proc_macro_value #136652)

After a line ending with ’', multiple lines are skipped.