Function proc_macro::is_available

pub fn is_available() -> bool

🔬 This is a nightly-only experimental API. (proc_macro_is_available #71436)

确定当前运行的程序是否可以访问 proc_macro。

Determines whether proc_macro has been made accessible to the currently running program.

proc_macro crate 仅用于在过程宏的实现中使用。 如果从过程宏外部调用，例如从构建脚本或单元测试或普通的 Rust 二进制文件调用，则此 crate 中的所有函数都会崩溃。

The proc_macro crate is only intended for use inside the implementation of procedural macros. All the functions in this crate panic if invoked from outside of a procedural macro, such as from a build script or unit test or ordinary Rust binary.

考虑到设计用于支持宏和非宏用例的 Rust 库，proc_macro::is_available() 提供了一种不惊慌的方法来检测使用 proc_macro API 所需的基础设施当前是否可用。 如果从程序宏内部调用，则返回 true，如果从任何其他二进制文件调用，则返回 false。

With consideration for Rust libraries that are designed to support both macro and non-macro use cases, proc_macro::is_available() provides a non-panicking way to detect whether the infrastructure required to use the API of proc_macro is presently available. Returns true if invoked from inside of a procedural macro, false if invoked from any other binary.