Struct proc_macro::Diagnostic

pub struct Diagnostic { /* private fields */ }

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

è¡¨ç¤ºè¯Šæ–­æ¶ˆæ¯å’Œå…³è”çš„å­æ¶ˆæ¯çš„ç»“æ„ä½“ã€‚

Implementations

impl Diagnostic

pub fn new<T: Into<String>>(level: Level, message: T) -> Diagnostic

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

ä½¿ç”¨ç»™å®šçš„ level å’Œ message åˆ›å»ºæ–°çš„è¯Šæ–­ã€‚

pub fn spanned<S, T>(spans: S, level: Level, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

ä½¿ç”¨ç»™å®šçš„ level å’Œ message æŒ‡å‘ç»™å®šçš„ spans é›†åˆ›å»ºæ–°çš„è¯Šæ–­ã€‚

pub fn span_error<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Error level, and the given spans and message.

pub fn error<T: Into<String>>(self, message: T) -> Diagnostic

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Error level, and the given message.

pub fn span_warning<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Warning level, and the given spans and message.

pub fn warning<T: Into<String>>(self, message: T) -> Diagnostic

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Warning level, and the given message.

pub fn span_note<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Note level, and the given spans and message.

pub fn note<T: Into<String>>(self, message: T) -> Diagnostic

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Note level, and the given message.

pub fn span_help<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Help level, and the given spans and message.

pub fn help<T: Into<String>>(self, message: T) -> Diagnostic

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Help level, and the given message.

pub fn level(&self) -> Level

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

è¿”å› self çš„è¯Šæ–­ levelã€‚

pub fn set_level(&mut self, level: Level)

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

å°† self ä¸­çš„çºåˆ«è®¾ç½®ä¸º levelã€‚

pub fn message(&self) -> &str

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

ä»¥ self è¿”å›æ¶ˆæ¯ã€‚

pub fn set_message<T: Into<String>>(&mut self, message: T)

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

å°† self ä¸­çš„æ¶ˆæ¯è®¾ç½®ä¸º messageã€‚

pub fn spans(&self) -> &[Span]

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

è¿”å› self ä¸­çš„ Spanã€‚

pub fn set_spans<S: MultiSpan>(&mut self, spans: S)

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

å°† self ä¸­çš„ Span è®¾ç½®ä¸º spansã€‚

pub fn children(&self) -> Children<'_>

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

è¿”å› self å­è¯Šæ–­çš„è¿­ä»£å™¨ã€‚

pub fn emit(self)

ğŸ”¬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

å‘å‡ºè¯Šæ–­ä¿¡æ¯ã€‚