Struct proc_macro::Group

pub struct Group(_);

A delimited token stream.

A Group internally contains a TokenStream which is surrounded by Delimiters.

Implementations

impl Group

pub fn new(delimiter: Delimiter, stream: TokenStream) -> Group

Creates a new Group with the given delimiter and token stream.

This constructor will set the span for this group to Span::call_site(). To change the span you can use the set_span method below.

pub fn delimiter(&self) -> Delimiter

Returns the delimiter of this Group

pub fn stream(&self) -> TokenStream

Returns the TokenStream of tokens that are delimited in this Group.

Note that the returned token stream does not include the delimiter returned above.

pub fn span(&self) -> Span

Returns the span for the delimiters of this token stream, spanning the entire Group.

pub fn span(&self) -> Span {
           ^^^^^^^

pub fn span_open(&self) -> Span

Returns the span pointing to the opening delimiter of this group.

pub fn span_open(&self) -> Span {
                ^

pub fn span_close(&self) -> Span

Returns the span pointing to the closing delimiter of this group.

pub fn span_close(&self) -> Span {
                       ^

pub fn set_span(&mut self, span: Span)

Configures the span for this Group’s delimiters, but not its internal tokens.

This method will not set the span of all the internal tokens spanned by this group, but rather it will only set the span of the delimiter tokens at the level of the Group.