Struct proc_macro::Span

pub struct Span(_);

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A region of source code, along with macro expansion information.

Methods

impl Span

pub fn call_site() -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The span of the invocation of the current procedural macro.

pub fn source_file(&self) -> SourceFile

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The original source file into which this span points.

pub fn start(&self) -> LineColumn

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Get the starting line/column in the source file for this span.

pub fn end(&self) -> LineColumn

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Get the ending line/column in the source file for this span.

pub fn join(&self, other: Span) -> Option<Span>

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new span encompassing self and other.

Returns None if self and other are from different files.

pub fn error<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn warning<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn note<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn help<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.