Struct proc_macro::TokenStream

pub struct TokenStream(_);

The main type provided by this crate, representing an abstract stream of tokens.

This is both the input and output of #[proc_macro_derive] definitions. Currently it's required to be a list of valid Rust items, but this restriction may be lifted in the future.

The API of this type is intentionally bare-bones, but it'll be expanded over time!

Methods

impl TokenStream

pub fn empty() -> TokenStream

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Returns an empty TokenStream.

pub fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Checks if this TokenStream is empty.