proc_macro

Module tracked

🔬This is a nightly-only experimental API. (proc_macro_tracked_path #99515)

Functionality for adding environment state to the build dependency info.

Functions

env_varExperimental

Retrieve an environment variable and add it to build dependency info. The build system executing the compiler will know that the variable was accessed during compilation, and will be able to rerun the build when the value of that variable changes. Besides the dependency tracking this function should be equivalent to env::var from the standard library, except that the argument must be UTF-8.

pathExperimental

Track a file or directory explicitly.