Function proc_macro::tracked_env::var

pub fn var<K: AsRef<OsStr> + AsRef<str>>(key: K) -> Result<String, VarError>

🔬 This is a nightly-only experimental API. (proc_macro_tracked_env #74690)

检索环境变量并将其添加到构建依赖项信息中。 执行编译器的构建系统将知道该变量在编译期间被访问，并且能够在该变量的值更改时重新运行构建。 除了依赖跟踪之外，这个函数应该等同于标准库中的 env::var，除了参数必须是 UTF-8。

Retrieve an environment variable and add it to build dependency info. Build system executing the compiler will know that the variable was accessed during compilation, and will be able to rerun the build when the value of that variable changes. Besides the dependency tracking this function should be equivalent to env::var from the standard library, except that the argument must be UTF-8.