Enum proc_macro::TokenTree

pub enum TokenTree {
    Group(Group),
    Ident(Ident),
    Punct(Punct),
    Literal(Literal),
}

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A single token or a delimited sequence of token trees (e.g. [1, (), ..]).

Variants
Group(Group)

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A token stream surrounded by bracket delimiters.

Ident(Ident)

🔬 This is a nightly-only experimental API. (proc_macro #38356)

An identifier.

Punct(Punct)

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A single punctuation character (+, ,, $, etc.).

Literal(Literal)

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A literal character ('a'), string ("hello"), number (2.3), etc.

Methods

impl TokenTree

pub fn span(&self) -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Returns the span of this tree, delegating to the span method of the contained token or a delimited stream.

pub fn set_span(&mut self, span: Span)

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Configures the span for only this token.

Note that if this token is a Group then this method will not configure the span of each of the internal tokens, this will simply delegate to the set_span method of each variant.