proc_macro::
Macro quote

pub macro quote($($t:tt)*) {
    ...
}

🔬This is a nightly-only experimental API. (proc_macro_quote #54722)

quote!(..) accepts arbitrary tokens and expands into a TokenStream describing the input. For example, quote!(a + b) will produce an expression, that, when evaluated, constructs the TokenStream [Ident("a"), Punct('+', Alone), Ident("b")].

Unquoting is done with $, and works by taking the single next ident as the unquoted term. To quote $ itself, use $$.