Struct proc_macro::Diagnostic

pub struct Diagnostic { /* private fields */ }

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

A structure representing a diagnostic message and associated children messages.

Implementations

impl Diagnostic

pub fn new<T: Into<String>>(level: Level, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new diagnostic with the given level and message.

pub fn spanned<S, T>(spans: S, level: Level, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new diagnostic with the given level and message pointing to the given set of spans.

pub fn span_error<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Error level, and the given spans and message.

pub fn error<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Error level, and the given message.

pub fn span_warning<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Warning level, and the given spans and message.

pub fn warning<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Warning level, and the given message.

pub fn span_note<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Note level, and the given spans and message.

pub fn note<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Note level, and the given message.

pub fn span_help<S, T>(self, spans: S, message: T) -> Diagnosticwhere S: MultiSpan, T: Into<String>,

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostics message to self with the Level::Help level, and the given spans and message.

pub fn help<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Adds a new child diagnostic message to self with the Level::Help level, and the given message.

pub fn level(&self) -> Level

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Returns the diagnostic level for self.

pub fn set_level(&mut self, level: Level)

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Sets the level in self to level.

pub fn message(&self) -> &str

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Returns the message in self.

pub fn set_message<T: Into<String>>(&mut self, message: T)

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Sets the message in self to message.

pub fn spans(&self) -> &[Span]

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Returns the Spans in self.

pub fn set_spans<S: MultiSpan>(&mut self, spans: S)

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Sets the Spans in self to spans.

pub fn children(&self) -> Children<'_>

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Returns an iterator over the children diagnostics of self.

pub fn emit(self)

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Emit the diagnostic.