Struct proc_macro::Punct

pub struct Punct(_);

A Punct is a single punctuation character such as +, - or #.

Multi-character operators like += are represented as two instances of Punct with different forms of Spacing returned.

Implementations

impl Punct

pub fn new(ch: char, spacing: Spacing) -> Punct

Creates a new Punct from the given character and spacing. The ch argument must be a valid punctuation character permitted by the language, otherwise the function will panic.

The returned Punct will have the default span of Span::call_site() which can be further configured with the set_span method below.

pub fn as_char(&self) -> char

Returns the value of this punctuation character as char.

pub fn spacing(&self) -> Spacing

Returns the spacing of this punctuation character, indicating whether it’s immediately followed by another Punct in the token stream, so they can potentially be combined into a multi-character operator (Joint), or it’s followed by some other token or whitespace (Alone) so the operator has certainly ended.

pub fn span(&self) -> Span

Returns the span for this punctuation character.

pub fn set_span(&mut self, span: Span)

Configure the span for this punctuation character.