Struct proc_macro::SourceFile

pub struct SourceFile { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The source file of a given Span.

Methods

impl SourceFile

pub fn path(&self) -> &FileName

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Get the path to this source file.

Note

If the code span associated with this SourceFile was generated by an external macro, this may not be an actual path on the filesystem. Use is_real to check.

Also note that even if is_real returns true, if --remap-path-prefix was passed on the command line, the path as given may not actually be valid.

pub fn is_real(&self) -> bool

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Returns true if this source file is a real source file, and not generated by an external macro's expansion.