Struct proc_macro::Span

pub struct Span(_);

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A region of source code, along with macro expansion information.

Methods

impl Span

pub fn def_site() -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

A span that resolves at the macro definition site.

pub fn call_site() -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The span of the invocation of the current procedural macro. Identifiers created with this span will be resolved as if they were written directly at the macro call location (call-site hygiene) and other code at the macro call site will be able to refer to them as well.

pub fn source_file(&self) -> SourceFile

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The original source file into which this span points.

pub fn parent(&self) -> Option<Span>

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The Span for the tokens in the previous macro expansion from which self was generated from, if any.

pub fn source(&self) -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

The span for the origin source code that self was generated from. If this Span wasn't generated from other macro expansions then the return value is the same as *self.

pub fn start(&self) -> LineColumn

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Get the starting line/column in the source file for this span.

pub fn end(&self) -> LineColumn

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Get the ending line/column in the source file for this span.

pub fn join(&self, other: Span) -> Option<Span>

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new span encompassing self and other.

Returns None if self and other are from different files.

pub fn resolved_at(&self, other: Span) -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Creates a new span with the same line/column information as self but that resolves symbols as though it were at other.

pub fn located_at(&self, other: Span) -> Span

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Creates a new span with the same name resolution behavior as self but with the line/column information of other.

pub fn eq(&self, other: &Span) -> bool

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Compares to spans to see if they're equal.

pub fn error<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn warning<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn note<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.

pub fn help<T: Into<String>>(self, message: T) -> Diagnostic

🔬 This is a nightly-only experimental API. (proc_macro #38356)

Create a new Diagnostic with the given message at the span self.