proc_macro

Struct Span
·

pub struct Span(/* private fields */);

A region of source code, along with macro expansion information.

Implementations

impl Span

pub fn def_site() -> Span

🔬This is a nightly-only experimental API. (proc_macro_def_site #54724)

A span that resolves at the macro definition site.

pub fn call_site() -> Span

The span of the invocation of the current procedural macro. Identifiers created with this span will be resolved as if they were written directly at the macro call location (call-site hygiene) and other code at the macro call site will be able to refer to them as well.

pub fn mixed_site() -> Span

A span that represents macro_rules hygiene, and sometimes resolves at the macro definition site (local variables, labels, $crate) and sometimes at the macro call site (everything else). The span location is taken from the call-site.

pub fn parent(&self) -> Option<Span>

🔬This is a nightly-only experimental API. (proc_macro_span #54725)

The Span for the tokens in the previous macro expansion from which self was generated from, if any.

pub fn source(&self) -> Span

🔬This is a nightly-only experimental API. (proc_macro_span #54725)

The span for the origin source code that self was generated from. If this Span wasn’t generated from other macro expansions then the return value is the same as *self.

pub fn byte_range(&self) -> Range<usize>

🔬This is a nightly-only experimental API. (proc_macro_span #54725)

Returns the span’s byte position range in the source file.

pub fn start(&self) -> Span

Creates an empty span pointing to directly before this span.

pub fn end(&self) -> Span

Creates an empty span pointing to directly after this span.

pub fn line(&self) -> usize

The one-indexed line of the source file where the span starts.

To obtain the line of the span’s end, use span.end().line().

pub fn column(&self) -> usize

The one-indexed column of the source file where the span starts.

To obtain the column of the span’s end, use span.end().column().

pub fn file(&self) -> String

The path to the source file in which this span occurs, for display purposes.

This might not correspond to a valid file system path. It might be remapped (e.g. "/src/lib.rs") or an artificial path (e.g. "<command line>").

pub fn local_file(&self) -> Option<PathBuf>

The path to the source file in which this span occurs on the local file system.

This is the actual path on disk. It is unaffected by path remapping.

This path should not be embedded in the output of the macro; prefer file() instead.

pub fn join(&self, other: Span) -> Option<Span>

🔬This is a nightly-only experimental API. (proc_macro_span #54725)

Creates a new span encompassing self and other.

Returns None if self and other are from different files.

pub fn resolved_at(&self, other: Span) -> Span

Creates a new span with the same line/column information as self but that resolves symbols as though it were at other.

pub fn located_at(&self, other: Span) -> Span

Creates a new span with the same name resolution behavior as self but with the line/column information of other.

pub fn eq(&self, other: &Span) -> bool

🔬This is a nightly-only experimental API. (proc_macro_span #54725)

Compares two spans to see if they’re equal.

pub fn source_text(&self) -> Option<String>

Returns the source text behind a span. This preserves the original source code, including spaces and comments. It only returns a result if the span corresponds to real source code.

Note: The observable result of a macro should only rely on the tokens and not on this source text. The result of this function is a best effort to be used for diagnostics only.

pub fn error<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new Diagnostic with the given message at the span self.

pub fn warning<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new Diagnostic with the given message at the span self.

pub fn note<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new Diagnostic with the given message at the span self.

pub fn help<T: Into<String>>(self, message: T) -> Diagnostic

🔬This is a nightly-only experimental API. (proc_macro_diagnostic #54140)

Creates a new Diagnostic with the given message at the span self.