Struct proc_macro::TokenStream

pub struct TokenStream { /* fields omitted */ }

The main type provided by this crate, representing an abstract stream of tokens.

This is both the input and output of #[proc_macro_derive] definitions. Currently it's required to be a list of valid Rust items, but this restriction may be lifted in the future.

The API of this type is intentionally bare-bones, but it'll be expanded over time!