Function proc_macro::tracked_env::var

pub fn var<K: AsRef<OsStr> + AsRef<str>>(key: K) -> Result<String, VarError>

🔬This is a nightly-only experimental API. (proc_macro_tracked_env #99515)

Retrieve an environment variable and add it to build dependency info. Build system executing the compiler will know that the variable was accessed during compilation, and will be able to rerun the build when the value of that variable changes. Besides the dependency tracking this function should be equivalent to env::var from the standard library, except that the argument must be UTF-8.